+++ packages/rath-client/public/locales/en-US.json
@@ -251,6 +251,14 @@
             "lastOpen": "Last opened",
             "firstOpen": "First created",
             "history": "Open Recent",
+            "history_time": {
+                "1d": "Today",
+                "1w": "Last weak",
+                "1mo": "Last month",
+                "3mo": "Last 3 months",
+                "6mo": "Last half a year",
+                "1yr": "Last year"
+            },
             "new": "New file",
             "sheet": "Sheet",
             "preview_parsed": "Preview",
