+++ packages/rath-client/public/locales/zh-CN.json
@@ -251,6 +251,14 @@
             "lastOpen": "上一次使用",
             "firstOpen": "创建时间",
             "history": "最近使用",
+            "history_time": {
+                "1d": "今天",
+                "1w": "过去一周",
+                "1mo": "过去一个月",
+                "3mo": "过去三个月",
+                "6mo": "过去半年",
+                "1yr": "一年内"
+            },
             "new": "新的文件",
             "sheet": "工作簿",
             "preview_parsed": "预览",
