+++ packages/rath-client/src/pages/dataSource/selection/history/index.tsx
@@ -23,6 +23,7 @@ const HistoryPanel: FC<Pick<IHistoryListProps, 'onDataLoaded' | 'onClose' | 'onL
                 onLoadingFailed={onLoadingFailed}
                 onClose={onClose}
                 search={search}
+                groupByPeriod
             />
         </>
     );
